modification of relation + Creation of HTML pages corrections of minor error
modification of relation to avoid problems when delete process.
modification of relation bills and money transfer to have the information of money transfer for each bill (1bill/moneyTransfer)
writing all the HTML code to verify all the method are working and i didn't make minor error.

TODO: Integration test, Acceptation test with the HTML, Delete all the classes relative to CreditCardInfo because a real company will call an exterior prestator to follow the juridiction on how to store credit card information.