changed in Contact and MoneyTransfer to avoid error when deleting user being placed in ContactedUser (FK2)
remove in the database the FK2 in Money Transfer and contact to User. switched for id in Contact in Money Transfer and deleted it Contact to replace it by string variables.

changed the html code thymleaf to see the result.

changed the test

implemented getContactById in contactRepository to use it in MoneyTransfer.

implemented getContactById in ContactService

Changed the controller